Refactor app structure, add landing page, and update TypeScript version
