+++ app/package.json
@@ -43,7 +43,7 @@
     "tailwindcss": "^3.4.17",
     "tailwindcss-animate": "^1.0.7",
     "tw-animate-css": "^1.2.8",
-    "typescript": "^5",
+    "typescript": "5.9.2",
     "vitest": "^3.1.3"
   }
 }
